relay canister messages to client session handlers
