+++ src/ic-websocket-gateway/src/canister_poller.rs
@@ -1,13 +1,19 @@
 use crate::{
     canister_methods::{
-        self, CanisterOpenMessageContent, CanisterOutputMessage, CanisterServiceMessage,
-        CanisterToClientMessage, CanisterWsCloseArguments, CanisterWsGetMessagesArguments,
-        ClientKey, WebsocketMessage,
+        self, CanisterOpenMessageContent, CanisterOutputCertifiedMessages, CanisterOutputMessage,
+        CanisterServiceMessage, CanisterToClientMessage, CanisterWsCloseArguments,
+        CanisterWsGetMessagesArguments, ClientKey, WebsocketMessage,
     },
-    events_analyzer::{Events, EventsCollectionType, EventsReference, IterationReference},
+    events_analyzer::{
+        Events, EventsCollectionType, EventsImpl, EventsReference, IterationReference,
+        MessageReference,
+    },
     manager::PollerState,
-    messages_demux::{MessagesDemux, CLIENTS_REGISTERED_IN_CDK},
-    metrics::canister_poller_metrics::{PollerEvents, PollerEventsMetrics},
+    messages_demux::CLIENTS_REGISTERED_IN_CDK,
+    metrics::canister_poller_metrics::{
+        IncomingCanisterMessageEvents, IncomingCanisterMessageEventsMetrics, PollerEvents,
+        PollerEventsMetrics,
+    },
 };
 use candid::{decode_one, Principal};
 use ic_agent::Agent;
@@ -26,7 +32,7 @@ use tokio::{
         RwLock,
     },
 };
-use tracing::{debug, error, info, span, trace, warn, Instrument, Level, Span};
+use tracing::{debug, error, info, span, trace, warn, Id, Instrument, Level, Span};
 
 // TODO: make sure this is always in sync with the CDK init parameter 'max_number_of_returned_messages'
 //       maybe get it once starting to poll the canister (?)
@@ -94,10 +100,16 @@ pub struct CanisterPoller {
     polling_iteration: u64,
     agent: Arc<Agent>,
     polling_interval_ms: u64,
+    analyzer_channel_tx: Sender<Box<dyn Events + Send>>,
 }
 
 impl CanisterPoller {
-    pub fn new(canister_id: Principal, poller_state: PollerState, agent: Arc<Agent>) -> Self {
+    pub fn new(
+        canister_id: Principal,
+        poller_state: PollerState,
+        agent: Arc<Agent>,
+        analyzer_channel_tx: Sender<Box<dyn Events + Send>>,
+    ) -> Self {
         Self {
             canister_id,
             poller_state,
@@ -109,6 +121,7 @@ impl CanisterPoller {
             agent,
             // TODO: let the canister be able to configure the polling interval
             polling_interval_ms: 100,
+            analyzer_channel_tx,
         }
     }
 
@@ -121,7 +134,7 @@ impl CanisterPoller {
         }
     }
 
-    async fn poll_canister(&self) -> Result<PollerEvents, String> {
+    async fn poll_canister(&mut self) -> Result<PollerEvents, String> {
         let polling_iteration_span =
             span!(Level::TRACE, "Polling Iteration", canister_id = %self.canister_id);
 
@@ -136,7 +149,7 @@ impl CanisterPoller {
         poller_events.metrics.set_start_polling();
 
         // get messages to be relayed to clients from canister (starting from 'message_nonce')
-        let mut certified_canister_output = canister_methods::ws_get_messages(
+        let certified_canister_output = canister_methods::ws_get_messages(
             &self.agent,
             &self.canister_id,
             CanisterWsGetMessagesArguments {
@@ -147,6 +160,8 @@ impl CanisterPoller {
         polling_iteration_span.in_scope(|| trace!("Polled canister for messages"));
 
         poller_events.metrics.set_received_messages();
+
+        let polling_interval_ms = self.polling_interval_ms;
 
         let number_of_returned_messages = certified_canister_output.messages.len();
         if number_of_returned_messages > 0 {
@@ -161,21 +176,13 @@ impl CanisterPoller {
                 });
 
                 poller_events.metrics.set_start_relaying_messages();
-                if false {
-                    return Err("Relaying messages failed".to_string());
+
+                if let Err(e) = self
+                    .relay_messages(certified_canister_output, polled_messages_span.id())
+                    .await
+                {
+                    return Err(format!("Relaying messages failed: {:?}", e));
                 }
-                // if let Err(e) = messages_demux
-                //     .write()
-                //     .await
-                //     .relay_messages(
-                //         certified_canister_output,
-                //         self.message_nonce,
-                //         polled_messages_span.id(),
-                //     )
-                //     .await
-                // {
-                //     return Err(e);
-                // }
                 polled_messages_span.in_scope(|| trace!("Relayed messages to connection handlers"));
                 Ok(())
             };
@@ -187,14 +194,116 @@ impl CanisterPoller {
                 relay_messages_async.await?;
                 polling_iteration_span.in_scope(|| trace!("Finished polling iteration"));
             } else {
-                tokio::join!(relay_messages_async, sleep(self.polling_interval_ms)).0?;
+                tokio::join!(relay_messages_async, sleep(polling_interval_ms)).0?;
             }
         } else {
             polling_iteration_span.in_scope(|| trace!("No messages polled"));
         }
+
+        // counting all polling iterations (instead of only the ones that return at least one canister message)
+        // this way we can tell for how many iterations the poller was "idle" before actually getting some messages from the canister
+        // this can help us in the future understanding whether the poller is polling too frequently or not
+        self.polling_iteration += 1;
         poller_events.metrics.set_finished_relaying_messages();
         Ok(poller_events)
     }
+
+    pub async fn relay_messages(
+        &mut self,
+        msgs: CanisterOutputCertifiedMessages,
+        polled_messages_span_id: Option<Id>,
+    ) -> Result<(), String> {
+        for canister_output_message in msgs.messages {
+            let canister_to_client_message = CanisterToClientMessage {
+                key: canister_output_message.key,
+                content: canister_output_message.content,
+                cert: msgs.cert.clone(),
+                tree: msgs.tree.clone(),
+            };
+            let mut incoming_canister_message_events = IncomingCanisterMessageEvents::new(
+                None,
+                EventsCollectionType::CanisterMessage,
+                IncomingCanisterMessageEventsMetrics::default(),
+            );
+            incoming_canister_message_events
+                .metrics
+                .set_start_relaying_message();
+
+            let last_message_nonce = get_nonce_from_message(&canister_to_client_message.key)?;
+            let message_key = MessageReference::new(self.canister_id, last_message_nonce.clone());
+            incoming_canister_message_events.reference =
+                Some(EventsReference::MessageReference(message_key));
+
+            // SAFETY:
+            // messages received from a client key that is not in the poller state are ignored
+            // this is safe to do because we the client session handler relayes the messages to the IC
+            // only after updating the poller state
+
+            // TODO: figure out if keeping references to a value in the poller state can cause deadlocks
+            if let Some((message_for_client_tx, client_connection_span)) = self
+                .poller_state
+                .get(&canister_output_message.client_key)
+                .as_deref()
+            {
+                let canister_message_span = span!(parent: client_connection_span, Level::TRACE, "canister_message", message_key = canister_to_client_message.key);
+                canister_message_span.follows_from(polled_messages_span_id.clone());
+                canister_message_span.in_scope(|| trace!("Received message from canister",));
+                self.relay_message(
+                    canister_to_client_message,
+                    message_for_client_tx,
+                    incoming_canister_message_events,
+                )
+                .instrument(canister_message_span)
+                .await;
+            }
+            // TODO: check without panicking
+            // assert_eq!(*message_nonce, last_message_nonce); // check that messages are relayed in increasing order
+
+            self.message_nonce = last_message_nonce + 1;
+        }
+        Ok(())
+    }
+
+    pub async fn relay_message(
+        &self,
+        canister_to_client_message: CanisterToClientMessage,
+        message_for_client_tx: &Sender<IcWsConnectionUpdate>,
+        mut incoming_canister_message_events: EventsImpl<IncomingCanisterMessageEventsMetrics>,
+    ) {
+        if let Err(e) = message_for_client_tx
+            .send(IcWsConnectionUpdate::Message((
+                canister_to_client_message,
+                Span::current(),
+            )))
+            .await
+        {
+            error!("Client's session terminated: {}", e);
+            incoming_canister_message_events
+                .metrics
+                .set_no_message_relayed();
+        } else {
+            incoming_canister_message_events
+                .metrics
+                .set_message_relayed();
+            trace!("Message relayed to connection handler");
+        }
+        self.analyzer_channel_tx
+            .send(Box::new(incoming_canister_message_events))
+            .await
+            .expect("analyzer's side of the channel dropped");
+    }
+}
+
+pub fn get_nonce_from_message(key: &String) -> Result<u64, String> {
+    if let Some(message_nonce_str) = key.split('_').last() {
+        let message_nonce = message_nonce_str
+            .parse()
+            .map_err(|e| format!("Could not parse nonce. Error: {:?}", e))?;
+        return Ok(message_nonce);
+    }
+    Err(String::from(
+        "Key in canister message is not formatted correctly",
+    ))
 }
 
 // let start_new_poller_span = span!(parent: &client_connection_span, Level::DEBUG, "start_new_poller_span", canister_id = %self.canister_id);
@@ -298,149 +407,84 @@ impl CanisterPoller {
 //     tokio::task::yield_now().await;
 // }
 
-pub fn filter_canister_messages<'a>(
-    messages: &'a mut Vec<CanisterOutputMessage>,
-    message_nonce: u64,
-    first_client_key: ClientKey,
-) {
-    if message_nonce == 0 {
-        // if the poller just started (message_nonce == 0), the canister might have already had other messages in the queue which we should not send to the clients
-        // therefore, starting from the last message polled, we relay the open message of type CanisterServiceMessage::OpenMessage for each connected client
-        // message_nonce has to be set to the nonce of the last open message pollled in this iteration so that in the next iteration we can poll from there
-        filter_messages_of_first_polling_iteration(messages, first_client_key);
-    }
-    // this is not the first polling iteration and therefore the poller queried the canister starting from the nonce of the last message of the previous polling iteration
-    // therefore, all the received messages are new and have to be relayed to the respective client handlers
-}
+// async fn signal_termination_and_cleanup(
+//     agent: &Arc<Agent>,
+//     poller_to_main_channel: &mut Sender<TerminationInfo>,
+//     canister_id: Principal,
+//     messages_demux: Rc<RefCell<MessagesDemux>>,
+//     e: String,
+// ) {
+//     // let the main task know that this poller will terminate due to a CDK error
+//     signal_poller_task_termination(
+//         poller_to_main_channel,
+//         TerminationInfo::CdkError(canister_id),
+//     )
+//     .await;
+//     // let each client connection handler task connected to this poller know that the poller will terminate
+//     // and thus they also have to close the WebSocket connection and terminate
+//     for (client_key, (client_channel_tx, _parent_span)) in messages_demux.borrow().client_channels()
+//     {
+//         call_ws_close_in_background(
+//             agent.clone(),
+//             canister_id,
+//             client_key.to_owned(),
+//             Span::current(),
+//         );
+//         if let Err(channel_err) = client_channel_tx
+//             .send(IcWsConnectionUpdate::Error(format!(
+//                 "Terminating poller task due to error: {}",
+//                 e
+//             )))
+//             .await
+//         {
+//             error!("Client's thread terminated: {}", channel_err);
+//         }
+//     }
+// }
 
-/// Finds the response to the open message of the client that started the poller.
-/// Returns all the messages following this message (inclusive).
-pub fn filter_messages_of_first_polling_iteration<'a>(
-    messages: &'a mut Vec<CanisterOutputMessage>,
-    first_client_key: ClientKey,
-) {
-    // the filter assumes that, if the response to the open message of the first client that connects after the gateway reboots is not (yet) present,
-    // the polled messages (which are all old) do not contain the result to a previous open message that the same client sent before the gateway rebooted
-
-    // if this is not the case, the filter will return all the messages from the result of the open message of the client which reconnects first
-    // however, this open message is old as it was sent by the client before the gateway rebooted and this is mistaken with
-    // the result of the new open message has not yet been pushed to the message queue of the canister (and thus not polled)
-
-    // this assumption is valid as the client is identified by its principal and a nonce which is generated for each new IC WS connection by the SDk
-    // therefore, if the same client reconnects, the client key will be different and the scenario mentioned above does not happen
-    let len_before_filter = messages.len();
-    messages.reverse();
-    let mut keep = true;
-    messages.retain(|canister_output_message| {
-        if keep {
-            let websocket_message: WebsocketMessage = from_slice(&canister_output_message.content)
-                .expect("content of canister_output_message is not of type WebsocketMessage");
-            if websocket_message.is_service_message {
-                let canister_service_message = decode_one(&websocket_message.content)
-                    .expect("content of websocket_message is not of type CanisterServiceMessage");
-                if let CanisterServiceMessage::OpenMessage(CanisterOpenMessageContent {
-                    client_key,
-                }) = canister_service_message
-                {
-                    if client_key == first_client_key {
-                        keep = false;
-                    }
-                }
-            }
-            return true;
-        }
-        false
-    });
-    // in case the response to the open message to the first client that connects is not found, all messages have to be discarded
-    // as they correspond to messages sent before the gateway rebooted
-    if keep == true {
-        messages.retain(|_m| false);
-    }
-    messages.reverse();
-    trace!(
-        "Filtered out {} polled messages",
-        len_before_filter - messages.len()
-    );
-}
-
-async fn signal_termination_and_cleanup(
-    agent: &Arc<Agent>,
-    poller_to_main_channel: &mut Sender<TerminationInfo>,
-    canister_id: Principal,
-    messages_demux: Rc<RefCell<MessagesDemux>>,
-    e: String,
-) {
-    // let the main task know that this poller will terminate due to a CDK error
-    signal_poller_task_termination(
-        poller_to_main_channel,
-        TerminationInfo::CdkError(canister_id),
-    )
-    .await;
-    // let each client connection handler task connected to this poller know that the poller will terminate
-    // and thus they also have to close the WebSocket connection and terminate
-    for (client_key, (client_channel_tx, _parent_span)) in messages_demux.borrow().client_channels()
-    {
-        call_ws_close_in_background(
-            agent.clone(),
-            canister_id,
-            client_key.to_owned(),
-            Span::current(),
-        );
-        if let Err(channel_err) = client_channel_tx
-            .send(IcWsConnectionUpdate::Error(format!(
-                "Terminating poller task due to error: {}",
-                e
-            )))
-            .await
-        {
-            error!("Client's thread terminated: {}", channel_err);
-        }
-    }
-}
-
-async fn signal_poller_task_termination(
-    channel: &mut Sender<TerminationInfo>,
-    info: TerminationInfo,
-) {
-    if let Err(e) = channel.send(info).await {
-        error!(
-            "Receiver has been dropped on the pollers connection manager's side. Error: {:?}",
-            e
-        );
-    }
-}
+// async fn signal_poller_task_termination(
+//     channel: &mut Sender<TerminationInfo>,
+//     info: TerminationInfo,
+// ) {
+//     if let Err(e) = channel.send(info).await {
+//         error!(
+//             "Receiver has been dropped on the pollers connection manager's side. Error: {:?}",
+//             e
+//         );
+//     }
+// }
 
 async fn sleep(millis: u64) {
     tokio::time::sleep(Duration::from_millis(millis)).await;
 }
 
-fn call_ws_close_in_background(
-    agent: Arc<Agent>,
-    canister_id: Principal,
-    client_key: ClientKey,
-    span: Span,
-) {
-    // close client connection on canister
-    // sending the request to the canister takes a few seconds
-    // therefore this is done in a separate task
-    // in order to not slow down the poller task
-    tokio::spawn(
-        async move {
-            debug!("Calling ws_close for client");
-            // TODO: figure out why it takes 10-30 seconds for the canister to close the connection
-            if let Err(e) = canister_methods::ws_close(
-                &agent,
-                &canister_id,
-                CanisterWsCloseArguments { client_key },
-            )
-            .await
-            {
-                error!("Calling ws_close on canister failed: {}", e);
-            } else {
-                debug!("Canister closed connection with client");
-            }
-            CLIENTS_REGISTERED_IN_CDK.fetch_sub(1, Ordering::SeqCst);
-        }
-        .instrument(span),
-    );
-}
+// fn call_ws_close_in_background(
+//     agent: Arc<Agent>,
+//     canister_id: Principal,
+//     client_key: ClientKey,
+//     span: Span,
+// ) {
+//     // close client connection on canister
+//     // sending the request to the canister takes a few seconds
+//     // therefore this is done in a separate task
+//     // in order to not slow down the poller task
+//     tokio::spawn(
+//         async move {
+//             debug!("Calling ws_close for client");
+//             // TODO: figure out why it takes 10-30 seconds for the canister to close the connection
+//             if let Err(e) = canister_methods::ws_close(
+//                 &agent,
+//                 &canister_id,
+//                 CanisterWsCloseArguments { client_key },
+//             )
+//             .await
+//             {
+//                 error!("Calling ws_close on canister failed: {}", e);
+//             } else {
+//                 debug!("Canister closed connection with client");
+//             }
+//             CLIENTS_REGISTERED_IN_CDK.fetch_sub(1, Ordering::SeqCst);
+//         }
+//         .instrument(span),
+//     );
+// }
